backend/requirements.txt
Django>=5.0,<6.0
gunicorn
whitenoise
django-cors-headers
mysqlclient
python-dotenv
djangorestframework
